pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
[project]
dependencies = [
 'ply',
 'python-dotenv',
]
name = "ringup_bsolis19"
version = "0.0.1"
authors = [
    { name="Ben Solis", email="benjairsolis19@gmail.com"},
]
description = "This program provides a data entry form for generating prices on goods"
readme = "README.md"
requires-python = ">=3.0"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
[project.urls]
"Homepage" = "https://github.com/bsolis19/Ring-Up"
"Bug Tracker" = "https://github.com/bsolis19/Ring-Up/issues"
